pyproject.toml
[project]
name = "plombery"
authors = [
    {name = "Luca Faggianelli", email = "luca@velvetlab.tech"},
]
description = "Python task scheduler with a user-friendly web UI"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
keywords = ["pipeline", "task-scheduler", "apscheduler", "orchestrator", "web-ui", "fastapi", "pydantic"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
  "APScheduler~=3.11.0",
  "apprise~=1.9.3",
  "authlib~=1.6.0",
  "fastapi~=0.115.12",
  "httpx~=0.28.1",
  "itsdangerous~=2.2.0",
  "jinja2~=3.1.6",
  "pydantic~=2.11.6",
  "pydantic-settings~=2.9.1",
  "python-socketio~=5.0.0",
  "python-dotenv~=1.1.0",
  "SQLAlchemy~=2.0.41",
  "uvicorn~=0.34.3",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "black~=25.1.0",
  "coverage~=7.9.1",
  "flake8~=7.2.0",
  "pytest~=8.4.0",
  "pytest_asyncio~=1.0.0",
  "pytest-env~=1.1.5",
]
docs = [
  "mkdocs-material>=9.1,<9.7",
]

[tool.setuptools.dynamic]
version = {attr="plombery._version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
# include = ["plombery*"]
# exclude = ["tests*"]

[tool.pytest.ini_options]
env = [
  "DATABASE_URL=sqlite:///:memory:"
]